Get database node ids or table metadata

Get database node ids or table metadata

get_tables(dbi, node, table = NULL, lang = c("et", "en"))

Arguments

dbi
TAI database id

node
node id

table
table id

lang
language. Default is Estonian 'et', 'en' English.

Value

a data.frame with id, type - types of node object l and t where l is s sublevel and t is a table, text - textual description

Details

Table metadata result has a title property and an array of variables. The variable object has four properties: code, text, elimination and time. The code and text properties are mandatory properties, while the elimination and time are optional. If time or elimination is not specified, the default value of “n” is used. The properties time and elimination could have either TRUE (yes) or FALSE (no) specified. If a variable has elimination set to “TRUE”, one can then omit selecting a value for that variable. There can only be one variable that has the time property set to “t” for a table. It also contains two lists. One that contains the codes for the all the values of which the variable can assume and one list of all the presentation text for the values.

References

http://www.scb.se/contentassets/79c32c72783a4f67b202ad3189f921b9/api-description.pdf